Final changes with resources

+++ shopping-cart/src/main/java/com/shopping/cart/dto/Coupon.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class Coupon {
     private String couponName;
     private Long appliedSku;
-    private Double discountPrice;
+    private BigDecimal discountPrice;
 }
